feat: price overide
